Do not consider URL as a unique field
This works around an apparent bug in FeedsProcessor.inc in which it
checks for unique targets, including the URL if the mapping indicates it
is unique, but it doesn't make sure there is any value at all before it
tries to fetch the entity ID.  So in our Zotero feed the URLs were all
blank so every search for a 'unique' entity brought back the first one
loaded.  But it therefore seems that URL is *not* unique for Zotero (or
it needs to be obtained correctly) so relying only on the guid to get
unique items seems to be the correct choice anyway.